ai-agent/backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
requests==2.32.4
python-dotenv==1.1.0 